requirements.txt
streamlit==1.2.0
bs4
requests
selenium
mecab-python3